+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.3.1
-
+- Improved Server logging
 - Improved Debug Watch
 
 ### 1.3.0